requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
sslcommerz-lib==1.0 